requirements.txt
requests[socks]
beautifulsoup4
python-dotenv
discord-webhook